Break xml parser into multiple modules
