+++ web_client/parser/index.js
@@ -1,0 +1,17 @@
+import constraints from './constraints';
+import convert from './convert';
+import defaultValue from './defaultValue';
+import group from './group';
+import panel from './panel';
+import parse from './parse';
+import widget from './widget';
+
+export {
+    constraints,
+    convert,
+    defaultValue,
+    group,
+    panel,
+    parse,
+    widget
+};
